LPS-96442 Inline JSModuleAdapter.
