--- modules/apps/frontend-js/frontend-js-loader-modules-extender/src/main/resources/com/liferay/frontend/js/loader/modules/extender/npm/model/package.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-
-<!--
-Copyright (c) 2000-present Liferay, Inc. All rights reserved.
-
-This library is free software; you can redistribute it and/or modify it under
-the terms of the GNU Lesser General Public License as published by the Free
-Software Foundation; either version 2.1 of the License, or (at your option)
-any later version.
-
-This library is distributed in the hope that it will be useful, but WITHOUT
-ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
-details.
--->
-
-<html>
-<body>
-<p>
-	Provides partial implementations for the interfaces in {@link
-	com.liferay.frontend.js.loader.modules.extender.npm.model}. These
-	implementations provide support for the fields and methods that are abstract
-	enough to be used by all existing implementations.
-</p>
-</body>
-</html>
